Issue a SIG_IGN before forking exec_commands.
Seems to prevent defunct sh processes from accumulating by allowing the OS to clear down the wrapping sh processes as soon as the complete.

The OS was retaining these processes in case we were going to come back and read the return status. We are not interested so we can use SIG_IGN to tell the OS that it can release these processes immediately.